requirements.txt
psutil
python-dotenv
requests
tzdata